refactor(EST-1682-BE): simplify error handling in Document class by using clear_error method
